requirements.txt
streamlit
pandas
faiss-cpu
sentence-transformers
google-generativeai
numpy
python-dotenv
